Create MovingStarDemo
I need to study the book examples more, but here is a star moving across the screen.  I need to figure out how to make the star respond to a mouse click (event handler), how to get the star to come in to random spots on the screen.  I would also like to figure out how to make it disappear.  And wouldn't it be nice if it says "20 points" for a half a second when the star disappears.