Updates (c) to NCG_v2 and MovingObect
I added a getter and setter method for the lblScore Label object in the main gui class.  I had to create the Label lblScore outside of the start code for the JavaFX application for this to work.  Had to make everything static so that I could use it in MovingObject class.  
Score is still not changing.