Create NCG_v2-update-e
Example Edit by Mark
This is what your inner classes should look more like. Changing the Shapes objects to ImageView objects should not affect the code. Let the GUI classes handle the event handling or find a way to make a method in the class to do so. Remeber that final implementation is ran by the main GUI class not 'target' classes. Will change MovingObject to abstract later.